Introduced new library to hold assume role credentials implementation
