+++ lib/ex_aws/config/auth_cache.ex
@@ -57,27 +57,11 @@ defmodule ExAws.Config.AuthCache do
       auth -> :ets.insert(ets, auth)
     end
 
-    assume_role_options = [
-      duration: credential_duration_seconds(expiration),
-      external_id: auth.external_id,
-    ]
-    role_session_name = auth.role_session_name || "default_session"
-
-    auth = if auth.role_arn do
-      {:ok, result} = auth.role_arn
-      |> ExAws.STS.assume_role(role_session_name, assume_role_options)
-      |> ExAws.Operation.perform(ExAws.Config.new(:sts))
-
-      IO.inspect auth
-
-      assumed_auth = %{
-        access_key_id: result.body.access_key_id,
-        secret_access_key: result.body.secret_access_key,
-        security_token: result.body.session_token,
-        expiration: result.body.expiration
-      }
+    auth = if auth[:role_arn] do
+      assumed_auth = ExAws.AssumeRoleCredentials.security_credentials(auth, expiration)
 
       :ets.insert(ets, {:awscli, assumed_auth})
+
       assumed_auth
     else
       auth
@@ -85,15 +69,6 @@ defmodule ExAws.Config.AuthCache do
 
     Process.send_after(self(), {:refresh_awscli_config, profile, expiration}, expiration)
     auth
-  end
-
-  defp credential_duration_seconds(expiration_ms) do
-    # assume_role accepts a duration between 900 and 3600 seconds
-    # We're adding a buffer to make sure the credentials live longer than
-    # the refresh interval.
-    {min, max, buffer} = {900, 3600, 5}
-    seconds = div(expiration_ms, 1000) + buffer
-    Enum.max([ Enum.min([max, seconds]), min ])
   end
 
   def refresh_config(config, ets) do
